Server projects: treat all warnings as errors
Fix all warnings across the server projects.

Note that ZipArchive is a bit ugly with continuing to downcast the ULONGLONG to a DWORD, but this is its existing behaviour
and it's considerably more effort to restructure the entire ZipArchive project to support 64-bit filesizes.

+++ Server/AIServer/MagicProcess.cpp
@@ -544,8 +544,8 @@ void CMagicProcess::ExecuteType4(int magicid, int sid, int tid, int data1, int d
 			pNpc->m_MagicType4[pType->BuffType - 1].byAmount = pType->Speed;
 			pNpc->m_MagicType4[pType->BuffType - 1].sDurationTime = pType->Duration;
 			pNpc->m_MagicType4[pType->BuffType - 1].fStartTime = TimeGet();
-			pNpc->m_fSpeed_1 = pNpc->m_fOldSpeed_1 * ((double) pType->Speed / 100);
-			pNpc->m_fSpeed_2 = pNpc->m_fOldSpeed_2 * ((double) pType->Speed / 100);
+			pNpc->m_fSpeed_1 = pNpc->m_fOldSpeed_1 * (pType->Speed / 100.0f);
+			pNpc->m_fSpeed_2 = pNpc->m_fOldSpeed_2 * (pType->Speed / 100.0f);
 			//TRACE(_T("executeType4 ,, speed1=%.2f, %.2f,, type=%d, cur=%.2f, %.2f\n"), pNpc->m_fOldSpeed_1, pNpc->m_fOldSpeed_2, pType->bSpeed, pNpc->m_fSpeed_1, pNpc->m_fSpeed_2);
 //			}
 			break;
@@ -760,9 +760,9 @@ short CMagicProcess::GetMagicDamage(int tid, int total_hit, int attribute, int d
 			bSign = false;
 		}
 
-		damage = (short) (total_hit - (0.7f * total_hit * total_r / 200));
+		damage = static_cast<short>(total_hit - (0.7f * total_hit * total_r / 200));
 		random = myrand(0, damage);
-		damage = (short) (0.7f * (total_hit - (0.9f * total_hit * total_r / 200))) + 0.2f * random;
+		damage = static_cast<short>((0.7f * (total_hit - (0.9f * total_hit * total_r / 200))) + 0.2f * random);
 //		damage = damage + (3 * righthand_damage);
 		damage = damage + righthand_damage;
 	}
@@ -1075,8 +1075,8 @@ void CMagicProcess::AreaAttackDamage(int magictype, int rx, int rz, int magicid,
 						pNpc->m_MagicType4[pType4->BuffType - 1].byAmount = pType4->Speed;
 						pNpc->m_MagicType4[pType4->BuffType - 1].sDurationTime = pType4->Duration;
 						pNpc->m_MagicType4[pType4->BuffType - 1].fStartTime = TimeGet();
-						pNpc->m_fSpeed_1 = pNpc->m_fOldSpeed_1 * ((double) pType4->Speed / 100);
-						pNpc->m_fSpeed_2 = pNpc->m_fOldSpeed_2 * ((double) pType4->Speed / 100);
+						pNpc->m_fSpeed_1 = pNpc->m_fOldSpeed_1 * (pType4->Speed / 100.0f);
+						pNpc->m_fSpeed_2 = pNpc->m_fOldSpeed_2 * (pType4->Speed / 100.0f);
 					//}
 						break;
